+++ src/main/java/com/example/goENC/models/compositeKey/ChoiceResponseId.java
@@ -1,8 +1,5 @@
 package com.example.goENC.models.compositeKey;
 
-import com.example.goENC.models.ChoiceAnswer;
-import com.example.goENC.models.Question;
-import com.example.goENC.models.Response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,7 +12,7 @@ import java.io.Serializable;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ChoiceResponseId implements Serializable {
-    private Integer questionId;
-    private Integer responseId;
-    private Integer choiceAnswerId;
+    private Long questionId;
+    private Long responseId;
+    private Long choiceAnswerId;
 }
